[#55] fix: Orders Entity 클래스명 변경에 따른 B2C 주문등록 에러 해결

+++ b2b/src/main/java/com/sparta/b2b/order/dto/response/OrderResponse.java
@@ -13,7 +13,7 @@ public record OrderResponse(
         OrderStatus orderStatus,
         DeliveryStatus deliveryStatus,
         Long totalPrice,
-        Long quantity,
+        Integer quantity,
         String b2cMemberName,
         Long b2bMemberId,
         String productName,
